+++ src/Mappers/Dominion/InfoMapper.php
@@ -80,7 +80,6 @@ class InfoMapper
             'employment' => $this->populationCalculator->getEmploymentPercentage($dominion),
             'networth' => $this->networthCalculator->getDominionNetworth($dominion),
             'prestige' => $dominion->prestige,
-            'infamy' => $dominion->infamy,
             'spy_resilience' => $dominion->spy_resilience,
             'wizard_resilience' => $dominion->wizard_resilience,
             'spy_mastery' => $dominion->spy_mastery,